+++ lib/src/async.dart
@@ -2,16 +2,14 @@ import 'dart:async';
 import 'package:angel_framework/angel_framework.dart';
 import 'package:angel_http_exception/angel_http_exception.dart';
 import 'package:matcher/matcher.dart';
+import 'context_aware.dart';
 
 /// Returns an [AngelMatcher] that uses an arbitrary function that returns
 /// true or false for the actual value.
 ///
 /// Analogous to the synchronous [predicate] matcher.
-///
-/// For example:
-///
-///     expect(v, predicate((x) => ((x % 2) == 0), "is even"))
-AngelMatcher predicateWithAngel(FutureOr<bool> Function(Object, Angel) f,
+AngelMatcher predicateWithAngel(
+        FutureOr<bool> Function(String, Object, Angel) f,
         [String description = 'satisfies function']) =>
     new _PredicateWithAngel(f, description);
 
@@ -19,26 +17,38 @@ AngelMatcher predicateWithAngel(FutureOr<bool> Function(Object, Angel) f,
 /// to the input.
 ///
 /// Use this to match values against configuration, injections, etc.
-AngelMatcher matchWithAngel(FutureOr<Matcher> Function(Object, Angel) f,
+AngelMatcher matchWithAngel(FutureOr<Matcher> Function(Object, Map, Angel) f,
         [String description = 'satisfies asynchronously created matcher']) =>
     new _MatchWithAngel(f, description);
 
 /// Calls [matchWithAngel] without the initial parameter.
+AngelMatcher matchWithAngelBinary(
+        FutureOr<Matcher> Function(Map context, Angel) f,
+        [String description = 'satisfies asynchronously created matcher']) =>
+    matchWithAngel((_, context, app) => f(context, app));
+
+/// Calls [matchWithAngel] without the initial two parameters.
 AngelMatcher matchWithAngelUnary(FutureOr<Matcher> Function(Angel) f,
         [String description = 'satisfies asynchronously created matcher']) =>
-    matchWithAngel((_, app) => f(app));
+    matchWithAngelBinary((_, app) => f(app));
+
+/// Calls [matchWithAngel] without any parameters.
+AngelMatcher matchWithAngelNullary(FutureOr<Matcher> Function() f,
+        [String description = 'satisfies asynchronously created matcher']) =>
+    matchWithAngelUnary((_) => f());
 
 /// Returns an [AngelMatcher] that represents [x].
 ///
 /// If [x] is an [AngelMatcher], then it is returned, unmodified.
 AngelMatcher wrapAngelMatcher(x) {
   if (x is AngelMatcher) return x;
-  return matchWithAngel((_, app) => wrapMatcher(x));
+  if (x is ContextAwareMatcher) return new _WrappedAngelMatcher(x);
+  return wrapAngelMatcher(wrapContextAwareMatcher(x));
 }
 
 /// Returns an [AngelMatcher] that asynchronously resolves a [feature], builds a [matcher], and executes it.
-AngelMatcher matchAsync(
-    FutureOr<Matcher> Function(Object) matcher, FutureOr Function() feature,
+AngelMatcher matchAsync(FutureOr<Matcher> Function(String, Object) matcher,
+    FutureOr Function() feature,
     [String description = 'satisfies asynchronously created matcher']) {
   return new _MatchAsync(matcher, feature, description);
 }
@@ -48,7 +58,7 @@ AngelMatcher matchAsync(
 AngelMatcher idExistsInService(String servicePath,
     {String idField: 'id', String description}) {
   return predicateWithAngel(
-    (item, app) async {
+    (key, item, app) async {
       try {
         var result = await app.service(servicePath)?.read(item);
         return result != null;
@@ -65,17 +75,34 @@ AngelMatcher idExistsInService(String servicePath,
 }
 
 /// An asynchronous [Matcher] that runs in the context of an [Angel] app.
-abstract class AngelMatcher extends Matcher {
-  Future<bool> matchesAsync(item, Map matchState, Angel app);
+abstract class AngelMatcher extends ContextAwareMatcher {
+  Future<bool> matchesWithAngel(
+      item, String key, Map context, Map matchState, Angel app);
 
   @override
-  bool matches(item, Map matchState) {
+  bool matchesWithContext(item, String key, Map context, Map matchState) {
     return true;
   }
 }
 
+class _WrappedAngelMatcher extends AngelMatcher {
+  final ContextAwareMatcher matcher;
+
+  _WrappedAngelMatcher(this.matcher);
+
+  @override
+  Description describe(Description description) =>
+      matcher.describe(description);
+
+  @override
+  Future<bool> matchesWithAngel(
+      item, String key, Map context, Map matchState, Angel app) async {
+    return matcher.matchesWithContext(item, key, context, matchState);
+  }
+}
+
 class _MatchWithAngel extends AngelMatcher {
-  final FutureOr<Matcher> Function(Object, Angel) f;
+  final FutureOr<Matcher> Function(Object, Map, Angel) f;
   final String description;
 
   _MatchWithAngel(this.f, this.description);
@@ -86,15 +113,16 @@ class _MatchWithAngel extends AngelMatcher {
       : description.add(this.description);
 
   @override
-  Future<bool> matchesAsync(item, Map matchState, Angel app) {
-    return new Future.sync(() => f(item, app)).then((result) {
+  Future<bool> matchesWithAngel(
+      item, String key, Map context, Map matchState, Angel app) {
+    return new Future.sync(() => f(item, context, app)).then((result) {
       return result.matches(item, matchState);
     });
   }
 }
 
 class _PredicateWithAngel extends AngelMatcher {
-  final FutureOr<bool> Function(Object, Angel) predicate;
+  final FutureOr<bool> Function(String, Object, Angel) predicate;
   final String description;
 
   _PredicateWithAngel(this.predicate, this.description);
@@ -105,13 +133,14 @@ class _PredicateWithAngel extends AngelMatcher {
       : description.add(this.description);
 
   @override
-  Future<bool> matchesAsync(item, Map matchState, Angel app) {
-    return new Future<bool>.sync(() => predicate(item, app));
+  Future<bool> matchesWithAngel(
+      item, String key, Map context, Map matchState, Angel app) {
+    return new Future<bool>.sync(() => predicate(key, item, app));
   }
 }
 
 class _MatchAsync extends AngelMatcher {
-  final FutureOr<Matcher> Function(Object) matcher;
+  final FutureOr<Matcher> Function(String, Object) matcher;
   final FutureOr Function() feature;
   final String description;
 
@@ -123,9 +152,11 @@ class _MatchAsync extends AngelMatcher {
       : description.add(this.description);
 
   @override
-  Future<bool> matchesAsync(item, Map matchState, Angel app) async {
+  Future<bool> matchesWithAngel(
+      item, String key, Map context, Map matchState, Angel app) async {
     var f = await feature();
-    var m = await matcher(f);
-    return m.matches(item, matchState);
+    var m = await matcher(key, f);
+    var c = wrapAngelMatcher(m);
+    return await c.matchesWithAngel(item, key, context, matchState, app);
   }
 }